+++ NEWS.md
@@ -1,5 +1,23 @@
 ## NEWS for the emmeans package
 
+emmeans 1.4.4
+---------------
+
+  * Fixed bug in `emmeans()` associated with non-factors such as `Date` (#162)
+  * Added `nesting.order` option to `emmip()` (#163)
+  * New `style` argument for `emmip()` allows plotting on a numeric scale
+  * More robust detection of response transformations (#166)
+  * Ensure `pwpp()` has tick marks on P-value axis (#167)
+  * Bug fix for `regrid()` for error when estimates exceed bounds
+  * Bug fix in auto-detecting nesting (#169) to make it less "enthusiastic"
+  * Fixes to formula operations needed because `formula.tools:::as.character.formula`
+    messes me up (thanks to Berwin Turloch, UWA, for alerting me)
+  * Making `dqrg()` more visible in the documentation (because it's often useful)
+  * Added more methods for `emm_list` objects, e.g. `rbind()` and `as.data.frame()`,
+    `as.list()`, and `as.emm_list()`
+  
+  
+
 emmeans 1.4.3.01
 ----------------